# feat(benchmark): add comprehensive performance benchmarking infrastructure