--- a/.github/workflows/benchmarks.yml
+++ b/.github/workflows/benchmarks.yml
@@ -44,12 +44,12 @@ jobs:
       if: runner.os == 'Linux'
       run: |
         sudo apt-get update
-        sudo apt-get install -y cmake ninja-build clang libbenchmark-dev libgtest-dev libfmt-dev libboost-all-dev
+        sudo apt-get install -y cmake ninja-build clang libbenchmark-dev libgtest-dev libfmt-dev libboost-all-dev libasio-dev
 
     - name: Install dependencies (macOS)
       if: runner.os == 'macOS'
       run: |
-        brew install ninja google-benchmark googletest fmt boost
+        brew install ninja google-benchmark googletest fmt boost asio
 
     - name: Set up compiler
       run: |