--- a/cmake/NetworkSystemIntegration.cmake
+++ b/cmake/NetworkSystemIntegration.cmake
@@ -10,7 +10,10 @@
 ##################################################
 function(setup_asio_integration target)
     if(USE_BOOST_ASIO)
-        target_include_directories(${target} PRIVATE ${Boost_INCLUDE_DIRS})
+        target_include_directories(${target}
+            PRIVATE
+                $<BUILD_INTERFACE:${Boost_INCLUDE_DIRS}>
+        )
         if(Boost_LIBRARIES)
             target_link_libraries(${target} PRIVATE ${Boost_LIBRARIES})
             message(STATUS "Configured ${target} with Boost.ASIO libraries")
@@ -19,7 +22,10 @@ function(setup_asio_integration target)
         endif()
         target_compile_definitions(${target} PRIVATE USE_BOOST_ASIO BOOST_ASIO_STANDALONE)
     elseif(ASIO_INCLUDE_DIR)
-        target_include_directories(${target} PRIVATE ${ASIO_INCLUDE_DIR})
+        target_include_directories(${target}
+            PRIVATE
+                $<BUILD_INTERFACE:${ASIO_INCLUDE_DIR}>
+        )
         target_compile_definitions(${target} PRIVATE ASIO_STANDALONE ASIO_NO_DEPRECATED)
         message(STATUS "Configured ${target} with standalone ASIO")
     endif()
@@ -39,7 +45,11 @@ function(setup_fmt_integration target)
         target_compile_definitions(${target} PUBLIC USE_FMT_LIBRARY)
         message(STATUS "Configured ${target} with fmt library (pkgconfig)")
     elseif(FMT_INCLUDE_DIR)
-        target_include_directories(${target} PUBLIC ${FMT_INCLUDE_DIR})
+        target_include_directories(${target}
+            PUBLIC
+                $<BUILD_INTERFACE:${FMT_INCLUDE_DIR}>
+                $<INSTALL_INTERFACE:include>
+        )
         if(FMT_LIBRARY)
             target_link_libraries(${target} PUBLIC ${FMT_LIBRARY})
             target_compile_definitions(${target} PUBLIC USE_FMT_LIBRARY)