--- a/include/network_system/integration/messaging_bridge.h
+++ b/include/network_system/integration/messaging_bridge.h
@@ -51,6 +51,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #endif
 
 #ifdef BUILD_WITH_THREAD_SYSTEM
+// Include fmt before thread_system headers since thread_system uses fmt formatters
+#  if defined(USE_FMT_LIBRARY)
+#    include <fmt/format.h>
+#  elif defined(USE_STD_FORMAT)
+#    include <format>
+#  endif
 #include <kcenon/thread/core/thread_pool.h>
 #endif