## Summary

Add Google Benchmark-based performance testing framework for network_system, establishing the foundation for Phase 0 baseline measurements and future performance regression detection.

### Key Changes

- **Benchmark Suite**: Comprehensive benchmarks for message throughput, connection management, and session handling
- **CMake Integration**: Added `NETWORK_BUILD_BENCHMARKS` option for conditional benchmark builds
- **CI/CD Workflow**: Automated benchmark execution on Ubuntu 22.04 and macOS platforms
- **Documentation**: Created `BASELINE.md` with target metrics and measurement methodology
- **Mock Infrastructure**: Implemented mock objects for isolated performance testing

### Benchmark Coverage

#### Message Operations (message_throughput_bench.cpp)
- Message creation, copy, move operations (64B to 64KB)
- Serialization and deserialization
- Queue push/pop operations
- Burst processing (10, 100, 1000 messages)
- Sustained throughput measurement
- Concurrent processing with multiple threads

#### Connection Management (connection_bench.cpp)
- Connection lifecycle (creation, establishment, destruction)
- Connection pool management (5, 10, 20, 50 connections)
- Send operations with various payload sizes
- Burst send operations
- State checking performance
- Concurrent operations with thread scaling

#### Session Management (session_bench.cpp)
- Session creation and destruction
- Data storage and retrieval
- Session lookup performance
- Batch session operations (10, 100, 1000 sessions)
- Cleanup operations
- Concurrent access patterns

### CI/CD Features

- Multi-platform testing (Ubuntu 22.04, macOS 13)
- JSON output for result archiving
- Baseline comparison support
- Artifact retention for 30 days
- Automated summary generation

### Files Added

- `.github/workflows/benchmarks.yml` - CI workflow configuration
- `benchmarks/CMakeLists.txt` - Benchmark build configuration
- `benchmarks/main_bench.cpp` - Benchmark entry point
- `benchmarks/message_throughput_bench.cpp` - Message performance tests
- `benchmarks/connection_bench.cpp` - Connection performance tests
- `benchmarks/session_bench.cpp` - Session performance tests
- `benchmarks/README.md` - Benchmark usage documentation
- `docs/BASELINE.md` - Performance baseline documentation

### Files Modified

- `CMakeLists.txt` - Added benchmark build option and subdirectory

## Test Plan

- [x] Benchmark code compiles successfully
- [x] CMake configuration accepts NETWORK_BUILD_BENCHMARKS option
- [ ] Run benchmarks locally to verify execution
- [ ] Verify CI workflow triggers on push
- [ ] Confirm benchmark results are uploaded as artifacts
- [ ] Review baseline documentation completeness

### Local Testing

```bash
cmake -B build -S . -DNETWORK_BUILD_BENCHMARKS=ON -DCMAKE_BUILD_TYPE=Release
cmake --build build
./build/benchmarks/network_benchmarks
```

### Expected Outcomes

- All benchmarks execute without crashes
- Results show consistent measurements across runs
- JSON output is valid and parseable
- CI workflow completes successfully on all platforms

## Performance Targets

Refer to `docs/BASELINE.md` for detailed performance targets:

- **Message Create (1KB)**: < 1μs (acceptable: < 10μs)
- **Connection Establishment**: < 100μs (acceptable: < 1ms)
- **Session Lookup**: < 100ns (acceptable: < 1μs)
- **Message Throughput**: > 100k msg/s (acceptable: > 50k msg/s)

## Future Work

- Install Google Benchmark and run actual measurements
- Update BASELINE.md with measured results
- Establish regression thresholds based on baseline
- Add real network socket benchmarks
- Implement SSL/TLS handshake benchmarks